move logic to OrNavButton comp
