+++ editor.planx.uk/src/@planx/components/shared/Preview/Card.tsx
@@ -3,18 +3,15 @@ import Button from "@mui/material/Button";
 import Container from "@mui/material/Container";
 import Fade from "@mui/material/Fade";
 import { styled, Theme, useTheme } from "@mui/material/styles";
-import { ComponentType as TYPES } from "@opensystemslab/planx-core/types";
 import { useAnalyticsTracking } from "pages/FlowEditor/lib/analytics/provider";
 import { useStore } from "pages/FlowEditor/lib/store";
 import React, { useEffect } from "react";
-import { ApplicationPath } from "types";
 
 import OrNavigationButton from "./OrNavigationButton";
 
 interface Props {
   children: React.ReactNode;
   isValid?: boolean;
-  isTestWarningWrapper?: boolean;
   handleSubmit?: (data?: any) => void;
 }
 
@@ -50,43 +47,16 @@ const Card: React.FC<Props> = ({
   children,
   isValid = true,
   handleSubmit,
-  isTestWarningWrapper,
   ...props
 }) => {
   const theme = useTheme();
-  const [path, visibleNode, breadcrumbs, flow] = useStore((state) => [
-    state.path,
-    state.currentCard,
-    state.breadcrumbs,
-    state.flow,
-  ]);
+  const [visibleNode] = useStore((state) => [state.currentCard]);
   const { track } = useAnalyticsTracking();
-
-  const defineNavigationType = (): OrNavigationType | undefined => {
-    // Check if we have a Send node in our breadcrumbs
-    //   This is a better/more immediate proxy for "submitted" in the frontend because actual send events that populate lowcal_sessions.submitted_at are queued via Hasura
-    const hasSent = Object.keys(breadcrumbs).some(
-      (breadcrumbNodeId: string) => flow[breadcrumbNodeId]?.type === TYPES.Send,
-    );
-
-    const showSaveResumeButton =
-      path === ApplicationPath.SaveAndReturn && handleSubmit && !hasSent;
-
-    if (showSaveResumeButton && !isTestWarningWrapper) {
-      return "save-resume";
-    }
-
-    if (!showSaveResumeButton && isTestWarningWrapper) {
-      return "navigate-to-published";
-    }
-  };
 
   useEffect(() => {
     // The Card component is only rendered when there's content the user will see
     if (visibleNode?.id) track(visibleNode?.id);
   }, []);
-
-  const orNavigationType = defineNavigationType();
 
   return (
     <Fade
@@ -118,9 +88,7 @@ const Card: React.FC<Props> = ({
                 Continue
               </Button>
             )}
-            {orNavigationType !== undefined && (
-              <OrNavigationButton type={orNavigationType} />
-            )}
+            <OrNavigationButton />
           </Box>
         </InnerContainer>
       </Container>
