+++ editor.planx.uk/src/pages/Preview/TestWarningPage.tsx
@@ -11,10 +11,7 @@ export const TestWarningPage = ({ children }: PropsWithChildren) => {
         children
       ) : (
         <Box width="100%">
-          <Card
-            handleSubmit={() => setHasAcknowledgedWarning(true)}
-            isTestWarningWrapper={true}
-          >
+          <Card handleSubmit={() => setHasAcknowledgedWarning(true)}>
             <CardHeader
               title="This is a test environment"
               description="This version of the service is unpublished and for testing only. Do not use it to submit real applications"
